refactor(krites/runtime): full structural overhaul — snafu errors, quality (#3216)
Split hnsw/put.rs (1170 LOC) into put.rs + graph.rs for focused
graph traversal vs insertion logic. Extract DRY helpers in sys.rs
(with_relation_write_lock, ok_rows) eliminating 5x repeated
lock-acquire-create-index-ok patterns. Add comprehensive //! module
docs covering HNSW algorithm, transaction semantics, semi-naive
evaluation, MinHash LSH banding, and imperative execution. Add
#[must_use] on NamedRows constructors and accessors, CallbackOp::as_str.
Fix #[allow] -> #[expect] on atomic_save.rs. All existing snafu error
propagation verified correct throughout.
