+++ crates/krites/src/runtime/hnsw/put.rs
@@ -1,15 +1,27 @@
 //! SessionTx methods for HNSW vector insertion.
+//!
+//! Implements the insertion half of the HNSW algorithm:
+//!
+//! - **`hnsw_put`**: public entry point that extracts vectors from a tuple,
+//!   enforces capacity limits, and dispatches to `hnsw_put_vector`.
+//!
+//! - **`hnsw_put_vector`**: inserts a single vector into the graph by selecting
+//!   a random level, traversing from the entry point down through layers, and
+//!   establishing bidirectional connections at each level.
+//!
+//! - **`hnsw_put_fresh_at_levels`**: initializes a new node at specified levels
+//!   when no entry point exists or when the node starts a new highest layer.
+//!
+//! Graph traversal and neighbor selection are in the sibling `graph` module.
 
 use std::cmp::{Reverse, max};
 
 use ordered_float::OrderedFloat;
 use priority_queue::PriorityQueue;
-use rustc_hash::FxHashSet;
 use tracing::warn;
 
 use super::idx_to_i64;
-use super::types::{CompoundKey, DEFAULT_VECTOR_CACHE_CAPACITY, HnswIndexManifest, VectorCache};
-use super::visited_pool::VisitedPool;
+use super::types::{DEFAULT_VECTOR_CACHE_CAPACITY, HnswIndexManifest, VectorCache};
 use crate::DataValue;
 use crate::data::expr::{Bytecode, eval_bytecode_pred};
 use crate::data::tuple::ENCODED_KEY_MIN_LEN;
@@ -21,6 +33,13 @@ use crate::runtime::transact::SessionTx;
 
 impl SessionTx<'_> {
     /// Insert a single vector into the HNSW index.
+    ///
+    /// Assigns a random level, traverses the graph from the entry point through
+    /// upper layers (greedy search with ef=1), then at the target layers uses
+    /// the construction beam width to find and establish neighbor connections.
+    ///
+    /// If the vector's hash matches an existing entry, insertion is skipped
+    /// (content-addressed deduplication).
     ///
     /// # Complexity
     ///
@@ -66,7 +85,6 @@ impl SessionTx<'_> {
             .next();
         if let Some(ep) = ep_res {
             let ep = ep?;
-            // SAFETY: `ep` comes from HNSW index scan which yields tuples with at least 1 element.
             let bottom_level = ep[0].get_int().ok_or_else(|| InvalidOperationSnafu {
                 op: "hnsw_read",
                 reason: "entry point bottom_level is not an integer".to_string(),
@@ -258,7 +276,6 @@ impl SessionTx<'_> {
                         clippy::cast_sign_loss,
                         reason = "degree is a small non-negative integer stored as f64"
                     )]
-                    // SAFETY: `target_self_val` is deserialized from HNSW metadata which always has at least 1 element.
                     let mut target_degree = target_self_val[0]
                         .get_float()
                         .ok_or_else(|| InvalidOperationSnafu {
@@ -306,353 +323,12 @@ impl SessionTx<'_> {
         }
         Ok(())
     }
-    /// Shrink neighbor connections when exceeding m_max.
-    ///
-    /// # Complexity
-    ///
-    /// O(m^2) where m is the maximum connections. Evaluates all pairwise distances
-    /// between neighbors for the heuristic selection.
-    fn hnsw_shrink_neighbour(
-        &mut self,
-        target_key: &CompoundKey,
-        m: usize,
-        level: i64,
-        manifest: &HnswIndexManifest,
-        idx_table: &RelationHandle,
-        orig_table: &RelationHandle,
-        vec_cache: &mut VectorCache,
-    ) -> Result<usize> {
-        vec_cache.ensure_key(target_key, orig_table, self)?;
-        let vec = vec_cache.get_key(target_key).clone();
-        let mut candidates = PriorityQueue::new();
-        for (neighbour_key, neighbour_dist) in
-            self.hnsw_get_neighbours(target_key, level, idx_table, false)?
-        {
-            candidates.push(neighbour_key, OrderedFloat(neighbour_dist));
-        }
-        let new_candidates = self.hnsw_select_neighbours_heuristic(
-            &vec,
-            &candidates,
-            m,
-            level,
-            manifest,
-            idx_table,
-            orig_table,
-            vec_cache,
-        )?;
-        let mut old_candidate_set = FxHashSet::default();
-        for (old, _) in &candidates {
-            old_candidate_set.insert(old.clone());
-        }
-        let mut new_candidate_set = FxHashSet::default();
-        for (new, _) in &new_candidates {
-            new_candidate_set.insert(new.clone());
-        }
-        let new_degree = new_candidates.len();
-        for (new, Reverse(OrderedFloat(new_dist))) in new_candidates {
-            if !old_candidate_set.contains(&new) {
-                let mut new_key = Vec::with_capacity(orig_table.metadata.keys.len() * 2 + 5);
-                let new_val = vec![
-                    DataValue::from(new_dist),
-                    DataValue::Null,
-                    DataValue::from(false),
-                ];
-                new_key.push(DataValue::from(level));
-                new_key.extend_from_slice(&target_key.0);
-                new_key.push(DataValue::from(idx_to_i64(target_key.1)));
-                new_key.push(DataValue::from(i64::from(target_key.2)));
-                new_key.extend_from_slice(&new.0);
-                new_key.push(DataValue::from(idx_to_i64(new.1)));
-                new_key.push(DataValue::from(i64::from(new.2)));
-                let new_key_bytes = idx_table.encode_key_for_store(&new_key, Default::default())?;
-                let new_val_bytes =
-                    idx_table.encode_val_only_for_store(&new_val, Default::default())?;
-                self.store_tx.put(&new_key_bytes, &new_val_bytes)?;
-            }
-        }
-        for (old, OrderedFloat(old_dist)) in candidates {
-            if !new_candidate_set.contains(&old) {
-                let mut old_key = Vec::with_capacity(orig_table.metadata.keys.len() * 2 + 5);
-                old_key.push(DataValue::from(level));
-                old_key.extend_from_slice(&target_key.0);
-                old_key.push(DataValue::from(idx_to_i64(target_key.1)));
-                old_key.push(DataValue::from(i64::from(target_key.2)));
-                old_key.extend_from_slice(&old.0);
-                old_key.push(DataValue::from(idx_to_i64(old.1)));
-                old_key.push(DataValue::from(i64::from(old.2)));
-                let old_key_bytes = idx_table.encode_key_for_store(&old_key, Default::default())?;
-                let old_existing_val = match self.store_tx.get(&old_key_bytes, false)? {
-                    Some(bytes) => bytes,
-                    None => {
-                        return Err(InvalidOperationSnafu {
-                            op: "hnsw_index",
-                            reason: "Indexed vector not found, this signifies a bug in the index implementation".to_string(),
-                        }.build().into())
-                    }
-                };
-                let old_existing_val: Vec<DataValue> = rmp_serde::from_slice(
-                    &old_existing_val[ENCODED_KEY_MIN_LEN..],
-                )
-                .map_err(|e| crate::error::InternalError::Runtime {
-                    source: InvalidOperationSnafu {
-                        op: "hnsw_index",
-                        reason: e.to_string(),
-                    }
-                    .build(),
-                })?;
-                if old_existing_val[2]
-                    .get_bool()
-                    .ok_or_else(|| InvalidOperationSnafu {
-                        op: "hnsw_index",
-                        reason: "deleted flag is not a boolean".to_string(),
-                    }.build())?
-                {
-                    self.store_tx.del(&old_key_bytes)?;
-                } else {
-                    let old_val = vec![
-                        DataValue::from(old_dist),
-                        DataValue::Null,
-                        DataValue::from(true),
-                    ];
-                    let old_val_bytes =
-                        idx_table.encode_val_only_for_store(&old_val, Default::default())?;
-                    self.store_tx.put(&old_key_bytes, &old_val_bytes)?;
-                }
-            }
-        }
 
-        Ok(new_degree)
-    }
-    /// Select neighbors using the HNSW heuristic (Algorithm 4 from the paper).
-    ///
-    /// # Complexity
-    ///
-    /// O(m^2 * ef) where m is max connections and ef is the beam width. Performs
-    /// pairwise distance checks between all candidates in the found set.
-    fn hnsw_select_neighbours_heuristic(
-        &self,
-        q: &Vector,
-        found: &PriorityQueue<CompoundKey, OrderedFloat<f64>>,
-        m: usize,
-        level: i64,
-        manifest: &HnswIndexManifest,
-        idx_table: &RelationHandle,
-        orig_table: &RelationHandle,
-        vec_cache: &mut VectorCache,
-    ) -> Result<PriorityQueue<CompoundKey, Reverse<OrderedFloat<f64>>>> {
-        let mut candidates = PriorityQueue::new();
-        let mut ret: PriorityQueue<CompoundKey, Reverse<OrderedFloat<_>>> = PriorityQueue::new();
-        let mut discarded: PriorityQueue<_, Reverse<OrderedFloat<_>>> = PriorityQueue::new();
-        for (item, dist) in found.iter() {
-            candidates.push(item.clone(), Reverse(*dist));
-        }
-        if manifest.extend_candidates {
-            for (item, _) in found.iter() {
-                for (neighbour_key, _) in self.hnsw_get_neighbours(item, level, idx_table, false)? {
-                    vec_cache.ensure_key(&neighbour_key, orig_table, self)?;
-                    let dist = vec_cache.v_dist(q, &neighbour_key)?;
-                    candidates.push(
-                        (neighbour_key.0, neighbour_key.1, neighbour_key.2),
-                        Reverse(OrderedFloat(dist)),
-                    );
-                }
-            }
-        }
-        while !candidates.is_empty() && ret.len() < m {
-            let (cand_key, Reverse(OrderedFloat(cand_dist_to_q))) =
-                candidates.pop().ok_or_else(|| InvalidOperationSnafu {
-                    op: "hnsw_select_neighbors",
-                    reason: "candidate queue unexpectedly empty".to_string(),
-                }.build())?;
-            let mut should_add = true;
-            for (existing, _) in ret.iter() {
-                vec_cache.ensure_key(&cand_key, orig_table, self)?;
-                vec_cache.ensure_key(existing, orig_table, self)?;
-                let dist_to_existing = vec_cache.k_dist(existing, &cand_key)?;
-                if dist_to_existing < cand_dist_to_q {
-                    should_add = false;
-                    break;
-                }
-            }
-            if should_add {
-                ret.push(cand_key, Reverse(OrderedFloat(cand_dist_to_q)));
-            } else if manifest.keep_pruned_connections {
-                discarded.push(cand_key, Reverse(OrderedFloat(cand_dist_to_q)));
-            }
-        }
-        if manifest.keep_pruned_connections {
-            while !discarded.is_empty() && ret.len() < m {
-                let (nearest_triple, Reverse(OrderedFloat(nearest_dist))) =
-                    discarded.pop().ok_or_else(|| InvalidOperationSnafu {
-                        op: "hnsw_select_neighbors",
-                        reason: "discarded queue unexpectedly empty".to_string(),
-                    }.build())?;
-                ret.push(nearest_triple, Reverse(OrderedFloat(nearest_dist)));
-            }
-        }
-        Ok(ret)
-    }
-    /// Search a single HNSW level, expanding the `found_nn` set.
-    ///
-    /// Delegates to [`hnsw_search_level_pooled`](Self::hnsw_search_level_pooled)
-    /// without a visited-list pool (fresh allocation per call).
-    ///
-    /// # Complexity
-    ///
-    /// O(ef * m) where ef is the beam width and m is max connections per node.
-    /// Visits at most ef nodes, checking m neighbors each.
-    pub(crate) fn hnsw_search_level(
-        &self,
-        q: &Vector,
-        ef: usize,
-        cur_level: i64,
-        orig_table: &RelationHandle,
-        idx_table: &RelationHandle,
-        found_nn: &mut PriorityQueue<CompoundKey, OrderedFloat<f64>>,
-        vec_cache: &mut VectorCache,
-    ) -> Result<()> {
-        self.hnsw_search_level_pooled(
-            q, ef, cur_level, orig_table, idx_table, found_nn, vec_cache, None,
-        )
-    }
-
-    /// Search a single HNSW level with optional visited-list pool.
-    ///
-    /// When a [`VisitedPool`] is provided, the visited set is acquired from the
-    /// pool and returned after use, eliminating per-search allocation.
-    ///
-    /// # Complexity
-    ///
-    /// O(ef * m) where ef is the beam width and m is max connections per node.
-    /// Space: O(ef) for the candidate priority queue plus O(ef) for visited set.
-    pub(crate) fn hnsw_search_level_pooled(
-        &self,
-        q: &Vector,
-        ef: usize,
-        cur_level: i64,
-        orig_table: &RelationHandle,
-        idx_table: &RelationHandle,
-        found_nn: &mut PriorityQueue<CompoundKey, OrderedFloat<f64>>,
-        vec_cache: &mut VectorCache,
-        visited_pool: Option<&VisitedPool>,
-    ) -> Result<()> {
-        let mut visited = match visited_pool {
-            Some(pool) => pool.acquire(),
-            None => FxHashSet::default(),
-        };
-        let mut candidates: PriorityQueue<CompoundKey, Reverse<OrderedFloat<f64>>> =
-            PriorityQueue::new();
-
-        for item in found_nn.iter() {
-            visited.insert(item.0.clone());
-            candidates.push(item.0.clone(), Reverse(*item.1));
-        }
-
-        while let Some((candidate, Reverse(OrderedFloat(candidate_dist)))) = candidates.pop() {
-            let (_, OrderedFloat(furtherest_dist)) =
-                found_nn.peek().ok_or_else(|| InvalidOperationSnafu {
-                    op: "hnsw_search_level",
-                    reason: "found_nn empty during level search".to_string(),
-                }.build())?;
-            let furtherest_dist = *furtherest_dist;
-            if candidate_dist > furtherest_dist {
-                break;
-            }
-            for (neighbour_key, _) in
-                self.hnsw_get_neighbours(&candidate, cur_level, idx_table, false)?
-            {
-                if visited.contains(&neighbour_key) {
-                    continue;
-                }
-                vec_cache.ensure_key(&neighbour_key, orig_table, self)?;
-                let neighbour_dist = vec_cache.v_dist(q, &neighbour_key)?;
-                let (_, OrderedFloat(cand_furtherest_dist)) =
-                    found_nn.peek().ok_or_else(|| InvalidOperationSnafu {
-                        op: "hnsw_search_level",
-                        reason: "found_nn empty during neighbor evaluation".to_string(),
-                    }.build())?;
-                if found_nn.len() < ef || neighbour_dist < *cand_furtherest_dist {
-                    candidates.push(neighbour_key.clone(), Reverse(OrderedFloat(neighbour_dist)));
-                    found_nn.push(neighbour_key.clone(), OrderedFloat(neighbour_dist));
-                    if found_nn.len() > ef {
-                        found_nn.pop();
-                    }
-                }
-                visited.insert(neighbour_key);
-            }
-        }
-
-        if let Some(pool) = visited_pool {
-            pool.release(visited);
-        }
-
-        Ok(())
-    }
-    /// Get neighbors of a node at a specific level.
-    ///
-    /// # Complexity
-    ///
-    /// O(m) where m is the number of neighbors (bounded by m_max).
-    pub(super) fn hnsw_get_neighbours<'b>(
-        &'b self,
-        cand_key: &'b CompoundKey,
-        level: i64,
-        idx_handle: &RelationHandle,
-        include_deleted: bool,
-    ) -> Result<impl Iterator<Item = (CompoundKey, f64)> + 'b> {
-        let mut start_tuple = Vec::with_capacity(cand_key.0.len() + 3);
-        start_tuple.push(DataValue::from(level));
-        start_tuple.extend_from_slice(&cand_key.0);
-        start_tuple.push(DataValue::from(idx_to_i64(cand_key.1)));
-        start_tuple.push(DataValue::from(i64::from(cand_key.2)));
-        let key_len = cand_key.0.len();
-        Ok(idx_handle
-            .scan_prefix(self, &start_tuple)
-            .filter_map(move |res| {
-                let tuple = res.ok()?;
-
-                #[expect(clippy::cast_sign_loss, reason = "HNSW indices are non-negative")]
-                #[expect(clippy::cast_possible_truncation, reason = "HNSW index fits in usize on all platforms")]
-                // INVARIANT: HNSW index tuples store int values at these positions
-                let key_idx = tuple[2 * key_len + 3]
-                    .get_int()
-                    .unwrap_or_else(|| unreachable!("HNSW neighbor index is not an integer")) as usize;
-                #[expect(
-                    clippy::cast_possible_truncation,
-                    reason = "HNSW subindex bounded by m_max (< i32::MAX)"
-                )]
-                let key_subidx = tuple[2 * key_len + 4]
-                    .get_int()
-                    .unwrap_or_else(|| unreachable!("HNSW neighbor subindex is not an integer")) as i32;
-                let key_tup = tuple[key_len + 3..2 * key_len + 3].to_vec();
-                if key_tup == cand_key.0 {
-                    None
-                } else {
-                    if include_deleted {
-                        return Some((
-                            (key_tup, key_idx, key_subidx),
-                            tuple[2 * key_len + 5]
-                                .get_float()
-                                .unwrap_or_else(|| unreachable!("HNSW neighbor distance is not a float")),
-                        ));
-                    }
-                    let is_deleted = tuple[2 * key_len + 7]
-                        .get_bool()
-                        .unwrap_or_else(|| unreachable!("HNSW deleted flag is not a boolean"));
-                    if is_deleted {
-                        None
-                    } else {
-                        Some((
-                            (key_tup, key_idx, key_subidx),
-                            tuple[2 * key_len + 5]
-                                .get_float()
-                                .unwrap_or_else(|| unreachable!("HNSW neighbor distance is not a float")),
-                        ))
-                    }
-                }
-            }))
-    }
     /// Insert a fresh vector at the specified levels without graph traversal.
+    ///
+    /// Used when inserting into an empty index or when the new node starts a
+    /// layer above the current entry point. Writes self-loop entries and a
+    /// canary entry for conflict detection.
     ///
     /// # Complexity
     ///
@@ -709,6 +385,7 @@ impl SessionTx<'_> {
         }
         Ok(())
     }
+
     /// Count vectors currently in the index by scanning the canary prefix.
     ///
     /// Canary entries (level = `DataValue::from(1)`) are written once per vector
@@ -723,6 +400,11 @@ impl SessionTx<'_> {
     }
 
     /// Public entry point for HNSW vector insertion.
+    ///
+    /// Extracts vector fields from `tuple` according to `manifest.vec_fields`,
+    /// enforces `max_vectors` capacity limits, and inserts each extracted vector
+    /// into the graph. Returns `true` if at least one vector was inserted, `false`
+    /// if the tuple was filtered out by the index filter.
     ///
     /// # Complexity
     ///
@@ -845,14 +527,14 @@ impl SessionTx<'_> {
                 Ok(t) => t,
                 Err(_) => continue,
             };
-            // Canary layout: [1, Null…, key_fields…, idx, subidx, Null…, Null, Null]
+            // Canary layout: [1, Null..., key_fields..., idx, subidx, Null..., Null, Null]
             // The tuple_key fields start at offset 1.
             let tuple_key: Vec<DataValue> = tuple.get(1..key_len + 1).unwrap_or_default().to_vec();
             if tuple_key.is_empty() {
                 continue;
             }
             match orig_table.get(self, &tuple_key) {
-                Ok(Some(_)) => {} // base row exists — consistent
+                Ok(Some(_)) => {} // base row exists -- consistent
                 Ok(None) => {
                     orphan_count = orphan_count.saturating_add(1);
                     warn!(
@@ -913,6 +595,7 @@ mod tests {
     // Insert `n` vectors into the index. Vector for id `i` is [i,i,i,i].
     fn insert_vectors(db: &DbInstance, n: usize) {
         for i in 0..n {
+            #[expect(clippy::cast_precision_loss, reason = "test fixture with small integers")]
             let val = i as f32;
             db.run_default(&format!(
                 "?[id, vec] <- [[{i}, vec([{val}, {val}, {val}, {val}])]] :put vectors {{}}"
@@ -920,10 +603,6 @@ mod tests {
             .unwrap();
         }
     }
-
-    // -----------------------------------------------------------------------
-    // hnsw_put: basic insert — vector is retrievable after insertion.
-    // -----------------------------------------------------------------------
 
     #[test]
     fn put_single_vector_is_retrievable() {
@@ -946,7 +625,6 @@ mod tests {
         assert_eq!(res.rows.len(), 10, "all 10 inserted vectors should be stored");
     }
 
-    // Duplicate insert (same key, same vector): idempotent — no duplicate rows.
     #[test]
     fn put_duplicate_key_is_idempotent() {
         let db = setup_db();
@@ -958,7 +636,6 @@ mod tests {
         assert_eq!(res.rows.len(), 1, "duplicate insert must not create extra rows");
     }
 
-    // Updating a vector (same key, different vector) replaces the old entry.
     #[test]
     fn put_updated_vector_replaces_old() {
         let db = setup_db();
@@ -966,7 +643,6 @@ mod tests {
             .unwrap();
         db.run_default("?[id, vec] <- [[7, vec([9.0, 9.0, 9.0, 9.0])]] :put vectors {}")
             .unwrap();
-        // Search near [9,9,9,9]: id=7 (updated) should be the closest.
         let res = db
             .run_default(
                 r#"?[id, dist] := ~vectors:idx{id | query: vec([9.0, 9.0, 9.0, 9.0]), k: 1, ef: 50, bind_distance: dist}"#,
@@ -978,7 +654,6 @@ mod tests {
         assert!(dist < 1e-6, "updated vector should be at distance ~0 from query");
     }
 
-    // Empty index: put nothing, search returns empty.
     #[test]
     fn put_empty_index_search_returns_nothing() {
         let db = setup_db();
@@ -990,24 +665,17 @@ mod tests {
         assert!(res.rows.is_empty(), "empty index must return no results");
     }
 
-    // -----------------------------------------------------------------------
-    // hnsw_search_level / hnsw_search_level_pooled: exercised via insertion
-    // (called during put for graph construction) and via KNN queries.
-    // -----------------------------------------------------------------------
-
-    // Search returns nearest neighbor in distance order after many inserts.
     #[test]
     fn search_level_returns_nearest_first() {
         let db = setup_db();
-        // Vectors [0,0,0,0] through [49,0,0,0] along a single axis.
         for i in 0..50 {
+            #[expect(clippy::cast_precision_loss, reason = "test fixture with small integers")]
             let val = i as f32;
             db.run_default(&format!(
                 "?[id, vec] <- [[{i}, vec([{val}, 0.0, 0.0, 0.0])]] :put vectors {{}}"
             ))
             .unwrap();
         }
-        // Query at [25,0,0,0]: id=25 should be first.
         let res = db
             .run_default(
                 r#"?[id, dist] := ~vectors:idx{id | query: vec([25.0, 0.0, 0.0, 0.0]), k: 5, ef: 50, bind_distance: dist} :order dist"#,
@@ -1018,7 +686,6 @@ mod tests {
         assert_eq!(first_id, 25, "nearest neighbor (id=25) must be returned first");
     }
 
-    // Distances in search results are non-decreasing (graph traversal is sound).
     #[test]
     fn search_level_results_non_decreasing_distance() {
         let db = setup_db();
@@ -1040,8 +707,6 @@ mod tests {
         }
     }
 
-    // Pool variant: search with ef=1 (greedy, single candidate) still finds the
-    // exact match when the vector is in the index.
     #[test]
     fn search_level_pooled_greedy_finds_exact_match() {
         let db = setup_db();
@@ -1051,25 +716,16 @@ mod tests {
                 r#"?[id] := ~vectors:idx{id | query: vec([10.0, 10.0, 10.0, 10.0]), k: 1, ef: 1, bind_distance: _dist}"#,
             )
             .unwrap();
-        // With ef=1 the graph traversal is very greedy — it may or may not find
-        // the exact match, but the search must complete without error.
         assert!(
             res.rows.len() <= 1,
             "greedy search must return at most k=1 results"
         );
     }
 
-    // -----------------------------------------------------------------------
-    // hnsw_get_neighbours: exercised via shrink-neighbour during dense insert.
-    // A large insert batch forces neighbour shrinking when m_max is exceeded.
-    // -----------------------------------------------------------------------
-
     #[test]
     fn get_neighbours_dense_insert_preserves_connectivity() {
-        // Insert many vectors — triggers hnsw_get_neighbours via shrink logic.
         let db = setup_db();
         insert_vectors(&db, 100);
-        // After 100 inserts the graph should still be searchable.
         let res = db
             .run_default(
                 r#"?[id, dist] := ~vectors:idx{id | query: vec([50.0, 50.0, 50.0, 50.0]), k: 5, ef: 50, bind_distance: dist}"#,
@@ -1079,7 +735,6 @@ mod tests {
             !res.rows.is_empty(),
             "graph must remain searchable after dense insert (neighbour shrink)"
         );
-        // codequality:ignore — test fixture with synthetic vector data, not production records
         assert!(
             res.rows.len() <= 5,
             "must return at most k=5 results, got {}",
@@ -1087,18 +742,6 @@ mod tests {
         );
     }
 
-    // -----------------------------------------------------------------------
-    // hnsw_check_consistency: the function is pub(crate) on SessionTx and is
-    // not yet wired to any Datalog command (dead_code).  The property it checks
-    // — that every canary entry has a corresponding base-relation row — is
-    // indirectly verified here: we insert vectors, confirm they are searchable,
-    // then remove one and confirm it is gone from both the base relation and the
-    // index.  A failing consistency check would manifest as search returning
-    // deleted rows, or inserts failing with corrupted-index errors.
-    // -----------------------------------------------------------------------
-
-    // After a clean batch of inserts, all vectors are present in the base
-    // relation (the condition hnsw_check_consistency validates).
     #[test]
     fn consistency_all_inserted_vectors_present_in_base_relation() {
         let db = setup_db();
@@ -1111,21 +754,17 @@ mod tests {
         );
     }
 
-    // After removing a vector, neither the base relation nor the index contains
-    // it — consistent state.
     #[test]
     fn consistency_deleted_vector_absent_from_index() {
         let db = setup_db();
         insert_vectors(&db, 10);
         db.run_default("?[id] <- [[5]] :rm vectors {}").unwrap();
 
-        // Base relation must not contain id=5.
         let base = db
             .run_default("?[id] := *vectors{id}, id = 5")
             .unwrap();
         assert!(base.rows.is_empty(), "deleted vector must be absent from base relation");
 
-        // Index must not return id=5 in search results.
         let search = db
             .run_default(
                 r#"?[id] := ~vectors:idx{id | query: vec([5.0, 5.0, 5.0, 5.0]), k: 10, ef: 50, bind_distance: _dist}"#,
@@ -1138,7 +777,6 @@ mod tests {
         );
     }
 
-    // Rebuild: drop and recreate the index — consistency is restored.
     #[test]
     fn consistency_after_index_rebuild() {
         let db = setup_db();
